+++ src/app/cart/cart.module.ts
@@ -1,15 +1,20 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+
+import { SharedModule } from './../shared/shared.module';
 import { CartListComponent } from './components/cart-list/cart-list.component';
+import { CartItemComponent } from './components/cart-item/cart-item.component';
 
 
 
 @NgModule({
   declarations: [
-    CartListComponent
+    CartListComponent,
+    CartItemComponent
   ],
   imports: [
     CommonModule,
+    SharedModule
   ],
   exports: [
     CartListComponent