+++ src/app/models/CartItem.model.ts
@@ -1,0 +1,6 @@
+import { Product } from './Product.model';
+
+export interface CartItem{
+    item: Product;
+    count: number;
+}