+++ src/app/shared/directives/highlight.directive.ts
@@ -1,0 +1,24 @@
+import { Directive, HostBinding, HostListener } from '@angular/core';
+
+@Directive({
+  selector: '[appHighlight]'
+})
+export class HighlightDirective {
+  color: string;
+  highlightColor = 'yellow';
+
+  @HostListener('mouseenter') onMouseEnter(): void {
+    this.color = this.highlightColor;
+  }
+
+  @HostListener('mouseleave') onMouseLeave(): void {
+    this.color = 'transparent';
+  }
+
+  @HostBinding('style.backgroundColor') get backgroundColor(): string {
+    return this.color;
+  }
+
+  constructor() { }
+
+}