+++ src/app/shared/shared.module.ts
@@ -1,12 +1,18 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { HighlightDirective } from './directives/highlight.directive';
 
 
 
 @NgModule({
-  declarations: [],
+  declarations: [
+    HighlightDirective
+  ],
   imports: [
     CommonModule
+  ],
+  exports: [
+    HighlightDirective
   ]
 })
 export class SharedModule { }